Adding the --enable-validating configuration option, which declares the   ACRO_VALIDATING and UTILIB_VALIDATING macros.  These replace the semantic   place of the DEBUGGING macro, which is deprecated.  Hence, all of these   changes...
The problem with DEBUGGING is that the --with-debugging configuration
option does _NOT_ turn this on ... and shouldn't.  DEBUGGING was really
intended to turn on code validation routines, and the name change reinforces
that.

NOTE: a side-effect of this change is that the macros DEBUGPR are always
defined now, which is what we want given that these are means of
controlling the verbosity of output.

( Links to Acro-related mail archives at http://software.sandia.gov/Acro/ )