We have these dependencies:
              /- exact/src <-- exact/test
              |
utilib/src <--|  pebbl/src <-- pebbl/test
              |     ^
              |     |
              \- pico/src  <-- pico/test

I modified the SUBDIRS directive in Makefile.am in the src and
test directories of pico, pebbl and exact to reflect these
dependencies.  So, for example, the Makefile in pico/src
goes into pebbl/src first to make sure it is up-to-date, and
the Makefile in pebbl/src goes into utilib/src first to make
sure it is up-to-date.  Updating a "src" directory also causes
it's common include directory to be updated. So, for example, if
a utilib header is changed, and then you do "make" in pico/src, utilib
and then pebbl will first be rebuilt, and the build in pico/src
will be using up-to-date utilib headers.