Misc bug fix in analysis.py.
Extended experiment.py to recognize the 'driver' attribute for the
        executable block.
Created a simple exact driver script, which runs the user command
with valgrind, and then greps for the summary statitistics.

NOTE: this sort of driver requires the user executable to find
the valgrind information.  We cannot automate this without requiring
a tighter coupling of EXACT ... which we've been trying not to do.

HOWEVER: we _could_ support a python module that a user script could
use to find 'standard' driver information (like valgrind info).


( Links to Acro-related mail archives at http://software.sandia.gov/Acro/ )