A stab at a python process manager that can appropriately supervised subprocesses to ensure that they run within a specified timelimit.  The SubprocessMngr class includes a lot of the functionality of the XyceRegression::Tools Perl module.
The main thing this doesn't do is figure out the right exit code
when running in parallel.  The problem is that that code is Xyce specific;
it depends on the detection of Xyce-specific output.  Xyce also
supports a code process monitoring feature: it only terminates if it looks
like the Xyce output file is not growing at a regular rate.

( Links to Acro-related mail archives at http://software.sandia.gov/Acro/ )