Rework of EXACT's executable path to point to the exact/src directory when testing exact itself.
Rework of memmon.c testing tool.  This used to compute the mapped
memory, but that does not correspond to the virtual memory size, which
this now grabs from /proc/$pid/status.

Added an EXACT test study that uses memmon and verifies that it's getting
a nonzero memory.

Revised the testfn.cpp command to sleep if a flag is passed.  This ensures
that memmon can assess it's peak memory usage.