Revisions of EXACT to remove the --preserve option and add the --force option (or the -f option).  This changes the execution semantics such that if an experiment directory or analysis file is found, then the experiment or analysis is only run if it is forced.  This has the impact of not allowing the specification of experiments or analyses with the same names (which was actually detected in example12.study.xml).
Misc fix in memmon.c

Augmented analyses to allow the specification of validation_methods and
comparison_fn's.  Added the ability for the validation analysis to
ignore the exit_status value.

Augmented example13 and example18 to use a validation method that
defaults to 'True' under cygwin.  This allows the tests to pass
in that context, even though some functionality is not supported for
MINGW builds.