Added option to support ordered sets.
Extended constructor rule semantics to allow for rules that 
specify the index into an ordered set.

Extended tutorial to illustrate the use of ordered sets.

Extended Set unit tests to include RangeSet unit tests.  Extended
and clarified RangeSet semantics.  These are now sets that contain
the elements start..end, but which can be revised with step parameters
and validation functions.