A major rework of the Config structure, to enable (a) plugins and (b) separate sections.
This also includes various updates of tests that have
identified some significant weaknesses of the Config
class:

  - Options must map to data of the same name
  - Option data must be stored in the Config object
        even for plugins
        (This could be avoided in the future)
  - Booleans cannot be given values 0 or 1 without
        making other 0/1 data look like booleans.

Despite these limitations, this is a pretty handy way
to manage global data.  The config options are now automatically
populated with the paths of commands needed to perform
IP format conversions.  Future plugins will add similar
data for IP solvers.  This ensures a more robust application
of these tools, as well as a natural manner for user-configuration
of these command paths.