Rework of the syntax for validation functions.  The old syntax was
  validate(value, i, j, model)

which I recently changed to

  validate(value, index, model)

However, this new syntax made coding awkward.  I've 
reverted to the old syntax, and changed Pyomo to give
'None' values for indices when validating a general value.  Thus,
we have

  validate(1.0, 1, 2, model)

will test if 1.0 is valid for index (1,2), but

  validate(1.0, None, None, model)

will test if 1.0 is valid for all possible indices.