Rework of memmon to avoid a bug found by Haixia.
I thought I had reworked the issues with the use of unsigned longs
before, but apparently not.  The new code does not use unsigned longs.

I also added more careful error checking.  When an error occurs getting the memory information, a check is made to confirm that the process has finished.  If so, the output is printed normally.  Otherwise, and error message is printed.