Such a little commit, but so important!
The 'nose' packages expects test functions to have a meaningful '__name__' attribute.
However, if you simply insert a lambda function, it's name is "<lambda>".

This fix creates a temporary object, to which we can set __name__,
and then we set the attribute with that temporary object!  Now, 'nose' is
happy!