A major rework of how the pyutilib.plugin package coordinates with plugins.  The framework now supports environment namespaces, which can be used to key configuration options.
Further, the Python 'logging' package is now used to support logging of 
the pyutili.plugin package (and associated packages).

All of this functionality may not work as advertised, but I'm committing this 
now because all of the existing tests pass.  Next, I'll work on additional tests
that exercise this new functionality.