Edits to get the ImportLoader working.
Various edits to add tests to improve coverage.

Setup skeleton of tests that will test the SimpleApplication
mechanism.  A goal of this is to more completely test the 
logger_config and env_config modules, which support convenient options
for simple applications.

Basic registration of the EggLoader has been tested, but not the core
functionality of this module.