This is a major revision of the pyutilib.plugin semantics.  The aim of this edit is to clarify the semantics of interfaces, and in particular interface namespaces relate to plugins.
Now, the namespace that an Interface subclass is declared
in defines the namespace where plugins will be registered.  Additionally,
a plugin will be registered in the namespace where it is declared.

For example:

  #
  # Declare Interface1 in namespace env1
  #
  PluginGlobals.push_env("env1")

  class Interface1(Interface):
        pass

  #
  # Declare Interface2 in namespace env2
  #
  PluginGlobals.push_env("env2")

  class Interface2(Interface):
        pass

  PluginGlobals.pop_env()

  #
  # Declare Plugin1 in namespace env3
  #
  PluginGlobals.push_env("env3")

  class Plugin1(Plugin):
    implements(Interface1)
    implements(Interface2)
    implements(Interface1,"env4")

  PluginGlobals.pop_env()


When Plugin1 is instantiated, it's services are registered
in the following environments:

    env1 for Interface1
    env2 for Interface2
    env4 for Interface1
    env3

The last registration is the default.  A plugin is always 
registered in the environment in which it was declared.