A variety of critical changes.  These were driven by bugs exposed in the Coopr tests, and they relate to basic functionality of the current plugin framework.
. Renamed ExtensionPoint.__getitem__ to __call__.  This provides for a more
natural syntax when asking for all of the extension points.  With getitem,
you need to pass in None to get that information, but with call you just
use ().

. Added a hook for the ManagedSingletonPlugin.  After much though, I decided
not to make this part of core.py.  That would have significantly extended
the requirements for this module (e.g. to include the core options code).

. Rework of the Plugin.__new__ method.  I simplified this to not explicitly
check if a PluginEnvironment is a subclass.  We're not using that
information, and this is not an expected derived class.

. Renamed Plugin.enable to Plugin._enable.  This allows the
ManagedPlugin class to implicitly override this value and provide a
configurable value for _enable.

. Added the ManagedPlugin and ManagedSingletonPlugin classes.  These classes
add a configuration option that can be used to enable/disable the plugin
within config files.  Various plugins were reworked to derive from this
class.

. Used super() in various constructor calls.  This seems to work,
though there is much to be said about the pitfalls of multiple inheritance,
even using super().  For now, I'm not going to worry too much about whether
multiple inheritance works correctly for the plugins class hierarchy.

. Reworked default arguments for various methods, when mutable types are passed in.  For example:

   def foo(arg=[]):

is a _dangerous_ declaration because if arg gets modified then it changes the
default value in subsequent calls to foo!  I knew about this, but I was
reminded of this problem this week.

. Added the plugin_services module, which provides simplified
interfaces to various singleton plugin services.


NOTE: I'm starting to see problems in the test_plugin/app/load tests.  The test_load2.py tests are
interferring with each other.  I haven't figured out the root cause of this interference, but I think that
failure/success in one test impacts the state of subsequent tests.