An extension of the PyUtilib plugin 'implements()' function.
I have added a 'inherit' option.  If this is set to True, then
the plugin implements the interface, and also inherits from this class.
This allows an interface class to define default methods for an
interface.

Although this is a generally useful utility, I did not make this 
the default behavior to help protect the user.  When this option is used,
the user can get into trouble when they mistype the name of the 
plugin method.  By making this mistake, the interface class API is used,
without any notification to the user of this behavior.  I expect that
this type of behavior could be quite difficult to track down.