Changing run_command() semantics.  By default, this assumes that a shell is used to execute the subprocess.
Although this differs from the Python defaults in the subprocess package,
the reality is that all of our uses of run_command in pyutilib, fast and
coopr make this assumption!  Further, the subprocess.Popen() function doesn't
clearly fail when passed a string that assumes a shell.  The output is often
an empty log file. {sigh}  Thus, tracking down run_command() with the shell
disabled (the former default) was quite difficult/time consuming.