Rework to enable termination of commands launched in a subshell.
This wasn't working earlier because I wasn't calling killpg(), which is
needed to kill the process group.

Unfortunately, this code currently causes an exception as Python is closing down:
  "Exception AttributeError: "'NoneType' object has no attribute 'error'" in <bound method Popen.__del__ ..."

This is due to a known bug in Python (see http://bugs.python.org/issue5099).  This is annoying, but it doesn't 
impact the execution of this code.