Follow expectation of least surprise.
When working with a set of ExtensionPoints, it is perhaps
counter-intuitive to sometimes __iter__ through them in one order,
and other times in another order.  It passed the "sniff test" as a
bug, while proving actually not to be.  This commit just makes it
more obvious that it's expected behavior.